requirements.txt
Django==4.2.9
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
Pillow==10.2.0
python-decouple==3.8
django-storages==1.14.2
